Entrega

+++ Philo/src/get_time.c
@@ -10,13 +10,13 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-# include "../include/philo.h"
+#include "../include/philo.h"
 
-uint64_t    ft_get_time(void)
+uint64_t	ft_get_time(void)
 {
-    struct timeval  time;
+	struct timeval	time;
 
-    if(gettimeofday(&time, NULL) == -1)
-        return ((uint64_t)-1);
-    return((uint64_t)time.tv_sec * 1000 + time.tv_usec / 1000);
+	if (gettimeofday(&time, NULL) == -1)
+		return (-1);
+	return ((uint64_t)time.tv_sec * 1000 + time.tv_usec / 1000);
 }
